+++ noodles-sam/src/writer/record/quality_scores.rs
@@ -3,7 +3,11 @@ use std::io::{self, Write};
 use super::MISSING;
 use crate::record::QualityScores;
 
-pub fn write_quality_scores<W>(writer: &mut W, quality_scores: &QualityScores) -> io::Result<()>
+pub fn write_quality_scores<W>(
+    writer: &mut W,
+    base_count: usize,
+    quality_scores: &QualityScores,
+) -> io::Result<()>
 where
     W: Write,
 {
@@ -11,11 +15,20 @@ where
 
     if quality_scores.is_empty() {
         writer.write_all(&[MISSING])?;
-    } else {
+    } else if quality_scores.len() == base_count {
         for &score in quality_scores.as_ref() {
             let n = u8::from(score) + MIN_VALUE;
             writer.write_all(&[n])?;
         }
+    } else {
+        return Err(io::Error::new(
+            io::ErrorKind::InvalidInput,
+            format!(
+                "sequence-quality scores length mismatch: expected {}, got {}",
+                base_count,
+                quality_scores.len()
+            ),
+        ));
     }
 
     Ok(())
@@ -27,14 +40,30 @@ mod tests {
 
     #[test]
     fn test_write_quality_scores() -> Result<(), Box<dyn std::error::Error>> {
-        let mut buf = Vec::new();
-        write_quality_scores(&mut buf, &QualityScores::default())?;
-        assert_eq!(buf, b"*");
+        fn t(
+            buf: &mut Vec<u8>,
+            base_count: usize,
+            quality_scores: &QualityScores,
+            expected: &[u8],
+        ) -> io::Result<()> {
+            buf.clear();
+            write_quality_scores(buf, base_count, quality_scores)?;
+            assert_eq!(buf, expected);
+            Ok(())
+        }
 
+        let mut buf = Vec::new();
+
+        t(&mut buf, 0, &QualityScores::default(), &[b'*'])?;
+        t(&mut buf, 4, &QualityScores::default(), &[b'*'])?;
+        t(&mut buf, 4, &"NDLS".parse()?, &[b'N', b'D', b'L', b'S'])?;
+
+        let quality_scores = "NDLS".parse()?;
         buf.clear();
-        let quality_scores: QualityScores = "NDLS".parse()?;
-        write_quality_scores(&mut buf, &quality_scores)?;
-        assert_eq!(buf, b"NDLS");
+        assert!(matches!(
+            write_quality_scores(&mut buf, 3, &quality_scores),
+            Err(e) if e.kind() == io::ErrorKind::InvalidInput
+        ));
 
         Ok(())
     }